requirements.txt
fastapi[standard]
python-dotenv

pydantic[email,timezone]

# Oracle
oracledb

#SqlAlchemy
sqlalchemy

alembic

bcrypt

python-jose[cryptography]


# Pour les tests
pytest